Fetch less often.
To be courteous to the owners of the server on which the renotation data
is kept, by default we now fetch only if 'all.dat' is not present.

So to force a refresh, delete 'all.dat' or move it out of the way.
Or if you're doing this through python, you can pass clobber=True to the
fetch() call.