A mess of reindexing.
This reworks indices for lightness and chroma slightly, adding black
values (at (0,0,0)) for lightness and shifting chroma indices to be
nonnegative.  Somewhere in there is a fix for a bug that mis-scaled
lightness values from 1.0 to 2.0.

This also adds NaN padding to the data array so that we don't have to
write as much special-casing of the boundary. The size cost shouldn't
matter too much thanks to gzip transport or something like that.

This also moves the window.onmousedown to the canvas's onmousedown,
since otherwise it and the "Render interpolated" button's onclick event
would jockey to run first.