Fix checking validity of sRGB vals near 0 or 255
Previously if the value of an sRGB coordinate on the 0 <= coord <= 1
scale was less than but close to 0, or greater than but close to 1,
it might end up rounding to a valid sRGB triple on the 0–255 scale but
the `color_valid` did not know this. This means a very few colours
(e.g. 7.5Y 5.00/8.0, 10.0P 1.00/8.0) which could have been shown even
with ‘Plot out-of-range colors’ unchecked were hidden from view.

This fixes the problem by making `color_valid` convert and round the
colour coordinates to the 0–255 scale used for display before checking
to see if they’re in range.